Explicitly set Markdownlint rules to enforce
Some of the rules enforce spaces instead of tabs, which isn’t the direction we’ve decided to go. The line length one is also annoying, especially when linking to outside repositories. This set of rules says to leave out some of the ones that don’t work for us.